Refactored src/services
